Refactor a few functions that are getting long
- Move all node logic for deploy into it's own function (getDeployNode is meant to be analogous to getDeployFsPath)
- Move registering commands to its own file in the commands folder
